pyproject.toml
[project]
name = "testmasterbsu"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.10"
dependencies = [
    "python-docx>=1.2.0",
    "lxml>=5.3.0",
    "pywin32>=311; sys_platform == 'win32'",
    "fastapi>=0.115.0",
    "python-multipart>=0.0.9",
    "uvicorn>=0.30.0",
    "pyinstaller>=6.18.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "bcrypt>=4.1.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib>=1.7.4",
    "pillow>=12.1.0",
    "email-validator>=2.0.0",
    "gunicorn>=23.0.0",
]
